+++ src/index.ts
@@ -89,7 +89,6 @@ import * as unfoldable from './Unfoldable'
 import * as validationT from './ValidationT'
 import * as witherable from './Witherable'
 import * as writer from './Writer'
-import * as writerT from './WriterT'
 export {
   /**
    * @since 2.0.0
@@ -438,9 +437,5 @@ export {
   /**
    * @since 2.0.0
    */
-  writer,
-  /**
-   * @since 2.4.0
-   */
-  writerT
+  writer
 }